Revert "06152020 update"
