+++ css/master.css
@@ -12,13 +12,9 @@
 --gray4: #8e9297;
 --white: #ffffff;
 
---main-bkg-color: var(--gray1);
---box-color: var(--gray2);
---post-color: var(--gray3);
-
---star-size: 35px;
---star-color: #fff;
---star-background: #fc0;
+  --star-size: 35px;
+  --star-color: #fff;
+  --star-background: #fc0;
 }
 html {
   color: var(--gray1);
@@ -93,7 +89,7 @@ hr {
 body {
   display: block;
   margin: 0;
-  background-color: var(--main-bkg-color);
+  background-color: teal;
   color: #fff;
   min-height: 100vh;
   font-family: sans-serif;
@@ -198,9 +194,7 @@ main {
  .logout-btn {
    margin-left: -10px;
  }
- .welcome-msg {
-   margin-right: 10px;
- }
+
   /*End Navigation */
   .header-login-btn {
     display: none;
@@ -413,51 +407,33 @@ main {
 }
 
 /* Home page Currently Reading */
-.currenty-reading {
-  background-color: var(--box-color);
-  width: 100%;
-  display: grid;
-  grid-template-columns: auto 50px;
-}
 .currenty-reading > .cur-text {
-  font-family: sans-serif;
-  font-weight: 100;
-  padding: 5px;
-}
-.currenty-reading .cur-text h1 {
-  font-size: 0.85em;
-}
-.currenty-reading .cur-text h3 {
-  font-size: 0.7em;
+  font-family: 'Patrick Hand SC', cursive;
 }
 .currenty-reading > .cur-text > * {
   margin: 0;
-  font-weight: 100;
-  letter-spacing: 0.8px;
 }
 .currenty-reading a {
   color: #fff;
 }
 .currenty-reading .cur-text h4 {
   text-decoration: underline;
-  font-size: 0.7em;
+  font-size: 15px;
   margin-top: 10px;
 }
 .currenty-reading .cur-text h5 {
-  font-size: 0.7em;
+  font-size: 18px;
+  margin-bottom: 15px;
 }
 .currenty-reading .cur-text h5 > span {
   text-decoration: underline;
   background-color: yellow;
   color: #000;
   padding: 0 3px;
-  font-weight: 500;
-  font-size: 1.4em;
-  cursor: pointer;
 }
 .currenty-reading > .book-cover-cur {
   display: flex;
-  width: 50px;
+  width: 100px;
   margin: 0 0 0 10px;
 }
 .currenty-reading > .book-cover-cur {
@@ -498,14 +474,20 @@ main {
     font-family: 'Patrick Hand SC', cursive;
     margin: 0;
   }
-
+  .currenty-reading > .book-cover-cur {
+    display: flex;
+    width: 100px;
+    margin: 0 0 0 10px;
+  }
 
 
   #main-view {
     width: 100%;
     display:flex;
   }
-
+  .currenty-reading {
+    margin: 50px auto;
+  }
   .cur-text {
     display: inline-block;
 
@@ -534,27 +516,17 @@ main {
   margin-top: 10px;
   text-decoration: none;
 }
-.updategoal-window {
+.upgoalbtn {
+  display: block;
+}
+.upgoalbtnhide {
   display: none;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
 }
-.update-goal-container {
-  padding: 10px;
-  background-color: var(--box-color);
-  border-radius: 15px;
-  margin: auto;
-  box-shadow: 0px 0px 60px #00ffff55;
-  width: 350px;
-}
-.pointer {
-  cursor: pointer;
+.updategoal {
+  display: none;
 }
 .updategoalshow {
-  display: flex;
+  display: block;
 }
 .uppgnum {
   font-size: 20px;
@@ -562,15 +534,12 @@ main {
   border-radius: 50px;
   border: none;
   outline: none;
-  width: 100%;
 }
 .updatepages {
   font-size: 22px;
   padding: 3px 15px;
   color: #fff;
   cursor: pointer;
-  margin: 10px 20px 20px;
-  width: -webkit-fill-available;
 }
 
 /* END: Home page Currently Reading */
@@ -594,7 +563,7 @@ main {
                        "pb-text pb-text pb-text"
                        ". . pb-button";
   padding: 10px;
-  background-color: var(--post-color);
+  background-color: var(--gray2);
 }
 .postbox__img {
   grid-area: pb-img;
@@ -609,9 +578,7 @@ main {
 }
 .postbox__text {
   grid-area: pb-text;
-  font-size: 12px;
-  padding: 15px;
-  border-radius: 22px;
+  font-size: 15px
 }
 .postbox__submit {
   grid-area: pb-button;
@@ -634,7 +601,7 @@ main {
                        "p-comm p-comm p-comm";
 
   padding: 10px;
-  background-color: var(--post-color);
+  background-color: var(--gray2);
 }
 
 .post__img {
@@ -1496,7 +1463,7 @@ main {
       width: 100%;
       height: 55px;
       box-shadow: 0 0 3px rgba(0, 0, 0, 0.2);
-      background-color: #000;
+      background-color: #202226;
       display: flex;
       overflow-x: auto;
       border-radius: 20px 20px 0 0;
@@ -1562,86 +1529,4 @@ main {
 }
 .spacer {
   flex-grow: 1;
-}
-
-
-/* New Blade UI */
-
-#bladeUI_grid {
-  display: grid;
-  grid-template-columns: 100%;
-  grid-auto-rows: auto;
-  grid-gap: 5px;
-}
-.readingNext,
-.discord_widget {
-  display: none;
-}
-@media (min-width: 900px) {
-  .stream {
-    grid-area:stream;
-    background-color: var(--box-color);
-    padding: 10px;
-    border-radius: 5px;
-  }
-  .currentlyReading{
-    grid-area:currently-reading;
-  }
-  .currenty-reading {
-    grid-template-columns: auto 100px;
-    border-radius: 5px;
-  }
-  .currenty-reading > .cur-text {
-    padding: 10px;
-  }
-  .currenty-reading .cur-text h1 {
-    font-size: 1.2em;
-  }
-  .currenty-reading .cur-text h3,
-  .currenty-reading .cur-text h5 {
-    margin-top: 5px;
-  }
-  .currenty-reading .cur-text h5,
-  .currenty-reading .cur-text h5 span {
-    font-size: 1.1em;
-  }
-  .currenty-reading > .book-cover-cur {
-    width: 100px;
-  }
-  .discord_widget {
-    display: block;
-    grid-area: discord_widget;
-    margin: auto;
-  }
-  .readingNext{
-    display: block;
-    grid-area:reading-next;
-    background-color: var(--box-color);
-    border-radius: 5px;
-  }
-  .readingNext h4 {
-    font-size: 1.3em;
-    margin: 15px;
-  }
-  #bladeUI_grid {
-    margin: 10px;
-    grid-gap: 10px;
-    grid-template-columns: auto 350px;
-    grid-template-areas: "stream currently-reading"
-                         "stream discord_widget"
-                         "stream reading-next"
-                         "stream .";
-  }
-}
-
-@media (min-width:1267px) {
-  #bladeUI_grid {
-    margin: 10px;
-    grid-gap: 20px;
-    grid-template-columns: auto 760px 350px auto;
-    grid-template-areas: ". stream currently-reading ."
-                         ". stream discord_widget ."
-                         ". stream reading-next ."
-                         ". stream . .";
-  }
 }
