+++ js/script.js
@@ -7,18 +7,14 @@ function openAcctMenu() {
   login.classList.add("open");
 }
 function updateGoal() {
-
-  var updateGoal = document.getElementById('updategoal-window');
-
-
-  updateGoal.classList.add("updategoalshow");
-
-}
-function closeUpdateGoal() {
-  var updateGoal = document.getElementById('updategoal-window');
-
-
-  updateGoal.classList.remove("updategoalshow");
+  var updategoalbtn = document.getElementById('updategoalbtn');
+  var pgnumGoal = document.getElementById('pgnumGoal');
+  var chapterGoal = document.getElementById('chapterGoal');
+  var customGoal = document.getElementById('customGoal');
+  updategoalbtn.classList.add("upgoalbtnhide");
+  pgnumGoal.classList.add("updategoalshow");
+  chapterGoal.classList.add("updategoalshow");
+  customGoal.classList.add("updategoalshow");
 }
 function openCommentArea(postID) {
   var commentFieldID = "comment" + postID;
@@ -33,4 +29,14 @@ document.querySelectorAll('textarea').forEach(el => {
     el.style.height = 'auto';
     el.style.height = (el.scrollHeight) + 'px';
   });
+});
+
+$(document).ready(function(){
+    // Check Radio-box
+    $(".rating input:radio").attr("checked", false);
+
+    $('.rating input').click(function () {
+        $(".rating span").removeClass('checked');
+        $(this).parent().addClass('checked');
+    });
 });
